% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_check.R
\name{check_hp}
\alias{check_hp}
\title{Calculate possible health points}
\usage{
check_hp(player_id, opponent_id, units_data, dist_matrix_data)
}
\arguments{
\item{player_id}{character ID of player}

\item{opponent_id}{character ID of opponent}

\item{units_data}{dataset contains information of all units in the game}

\item{dist_matrix_data}{distance matrix}
}
\value{
Possible number of health points in case of attack but without any boosters and discounts
}
\description{
Calculate possible health points for information purposes
}
\examples{

check_hp("")
}
