% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_matrix.R
\docType{data}
\name{dist_matrix}
\alias{dist_matrix}
\title{Distance matrix}
\format{An object of class \code{matrix} with 357 rows and 357 columns.}
\usage{
dist_matrix
}
\description{
Distance matrix for hex map with 357 hexs (16 x 23)
}
\examples{
dist_matrix()
}
\keyword{datasets}
