% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_reduction.R
\name{distance_reduction}
\alias{distance_reduction}
\title{Distance reduction value}
\usage{
distance_reduction(range_min, range_max, accuracy, distance)
}
\arguments{
\item{range_min}{minimum range of the unit}

\item{range_max}{maximum range of the unit}

\item{accuracy}{accuracy of the unit}

\item{distance}{distance between unit and target}
}
\value{
distance reduction value from [0;1] interval
}
\description{
Calculate distance reduction value due to range and accuracy of the unit and distance from the target.
}
\examples{
distance_reduction(range_min = 1, range_max = 5, accuracy = 6, distance = 4)
}
