% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_hexs.R
\name{draw_hexs}
\alias{draw_hexs}
\title{Draw hexs}
\usage{
draw_hexs(player1, player2)
}
\arguments{
\item{player1}{drawned units for player 1}

\item{player2}{drawned units for player 2}
}
\value{
Dataset of drawned units with hexs
}
\description{
Draw hexs from left and right side of the map for drawned units
}
\examples{

player1 <- draw_units(price_limit = 50, units_data = units)
player2 <- draw_units(price_limit = 50, units_data = units)

draw_hexs(player1, player2)
}
