% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_units.R
\name{draw_units}
\alias{draw_units}
\title{Draw units}
\usage{
draw_units(price_limit = 50, units_data)
}
\arguments{
\item{price_limit}{limit of price within units can be sampled}

\item{units_data}{dataset contains information of available units}
}
\value{
Dataset of drawned units according to price limit
}
\description{
Draw units according to indicated price
}
\examples{

player1 <- draw_units(price_limit = 50, units_data = units)
}
