% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_check.R
\name{hp_check}
\alias{hp_check}
\title{Calculate possible health points}
\usage{
hp_check(player_id, opponent_id, units_data, dist_matrix_data)
}
\arguments{
\item{player_id}{character ID of player}

\item{opponent_id}{character ID of opponent}

\item{units_data}{dataset contains information of all units in the game}

\item{dist_matrix_data}{distance matrix}
}
\value{
Possible number of health points in case of attack but without any boosters and discounts
}
\description{
Calculate possible health points for information purposes
}
\examples{

set.seed(123)
p1 <- draw_units(price_limit = 50, units_data = units)
p2 <- draw_units(price_limit = 50, units_data = units)
units_game <- draw_hexs(p1, p2)

hp_check(player_id = "ID9", opponent_id = "ID14",
units_data = units_game, dist_matrix_data = dist_matrix)
}
