% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Calculate beta coefficients for a straight line connecting two points. This function is used to calculate strength of attack with distance redcution.}
\usage{
linear_model(a, b)
}
\arguments{
\item{a}{first point - numeric vector of length of 2}

\item{b}{second point - numeric vector of length of 2}
}
\value{
Beta coefficients for line which connects points a and b
}
\description{
Calculate beta coefficients for a straight line connecting two points. This function is used to calculate strength of attack with distance redcution.
}
\examples{
linear_model(c(0, 10), c(10, 1))
}
