% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{units}
\alias{units}
\title{Units data}
\format{A data frame with 5 rows and 11 variables}
\usage{
units
}
\description{
Examples of units in game. The variables are as follows:
}
\details{
\describe{
\item{name}{character; name of unit}
\item{hp}{numeric; initial health points}
\item{mobil}{numeric; mobility points}
\item{attack}{numeric; attack points}
\item{defence}{numeric; defence points}
\item{range_min}{numeric; minimum range of unit}
\item{range_max}{numeric; maximum range of unit}
\item{price}{numeric; price of unit}
\item{acc}{numeric; accuracy points}
\item{acc_min}{numeric; missing shot chance}
\item{acc_max}{numeric; super shot chance}
}
}
\keyword{datasets}
