% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexs_mobility.R
\name{hexs_mobility}
\alias{hexs_mobility}
\title{Mobility hexs for unit}
\usage{
hexs_mobility(hex_id, mobility = NULL, units_data, dist_matrix_data)
}
\arguments{
\item{hex_id}{id of hex where there is a unit}

\item{mobility}{mobility of unit, default is NULL and then mobility is taken from units_data}

\item{units_data}{dataset which contains information about units with its hexs}

\item{dist_matrix_data}{distance matrix}
}
\value{
Vector with IDs of hexs where unit can move.
}
\description{
Return a vector of hexs which are empty, so the unit can move on them.
}
\examples{
hexs_mobility("ID8", units_data = units_game, dist_matrix_data = dist_matrix)
}
